pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raw-llm"
version = "1.0.0"
description = "The simplest way to context engineer. Minimal streaming CLI clients for Claude and Gemini."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Rodolfo Villaruz", email = "rodolfo@yes.ph"}
]
keywords = ["llm", "claude", "gemini", "cli", "streaming", "context-engineering"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "anthropic>=0.25.0",
    "google-genai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "pylint>=2.17.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest>=7.3.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.scripts]
raw-claude = "raw_llm.claude:main"
raw-gemini = "raw_llm.gemini:main"

[project.urls]
Homepage = "https://github.com/rodolfovillaruz/simple"
"Bug Tracker" = "https://github.com/rodolfovillaruz/simple/issues"
Repository = "https://github.com/rodolfovillaruz/simple.git"

[tool.setuptools]
packages = ["raw_llm"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pylint.messages_control]
disable = ["C0111", "R0903"]
